各种集成的实现

+++ main.py
@@ -5,28 +5,33 @@
 @author  : leafw
 """
 import requests
-
-import pojo
-from llm import kimi, deepseek
-from prompt_template import paper_questions
-import streamlit as st
-import pandas as pd
-from streamlit_option_menu import option_menu
 import os
 import utils
-from pojo import ArxivData
+from urllib.parse import urlparse
+import streamlit as st
+from streamlit_option_menu import option_menu
+import pandas as pd
+from prompt_template import paper_questions
+from pojo import load_Arxiv_data, ArxivData
 from api import arxiv_client
+from llm.model import OpenAiLlm, KimiLlm, OllamaLlm
+from llm.agent import TranslaterAgent, PaperAnswerAgent
+
+# 是否使用Kimi
+use_kimi = False
+# 是否使用ollama
+use_ollama = False
 
 data_dir = './data'
-
-
-def arxiv_dir_path(arxiv_id: str):
-    return data_dir + os.sep + arxiv_id
+md_template_path = 'md_template.md'
+current_llm = KimiLlm() if use_kimi else OllamaLlm('qwen') if use_ollama else OpenAiLlm('deepseek')
+trans_agent = TranslaterAgent(llm=current_llm)
+paper_answer_agent = PaperAnswerAgent(llm=current_llm)
 
 
 def get_data_from_arxiv_id(arxiv_id: str) -> ArxivData:
-    json_path = arxiv_dir_path(arxiv_id) + os.sep + arxiv_id + '.json'
-    return pojo.load_Arxiv_data(json_path)
+    json_path = utils.arxiv_dir_path(arxiv_id) + os.sep + arxiv_id + '.json'
+    return load_Arxiv_data(json_path)
 
 
 def download(url):
@@ -36,10 +41,14 @@ def download(url):
 
     arxiv_id = url[last + 1:]
     # 每个论文放在data/arxiv_id文件夹下
-    directory_path = arxiv_dir_path(arxiv_id)
+    directory_path = utils.arxiv_dir_path(arxiv_id)
     utils.ensure_directory_exists(directory_path)
 
     file_path = directory_path + os.sep + arxiv_id + '.pdf'
+
+    if os.path.exists(file_path):
+        print('文件已经存在')
+        return file_path, arxiv_id
 
     if response.status_code == 200:
         with open(file_path, 'wb') as file:
@@ -51,8 +60,19 @@ def download(url):
 
 
 def parse_home(url: str) -> dict:
-    arxiv_data = arxiv_client.search_by_url(url)
+    parsed_url = urlparse(url)
+    arxiv_id = parsed_url.path.split('/')[-1]
 
+    arxiv_data = get_data_from_arxiv_id(arxiv_id)
+    if arxiv_data:
+        return {
+            "title": arxiv_data.title,
+            "abstract": arxiv_data.abstract,
+            "file_path": arxiv_data.file_path,
+            "arxiv_id": arxiv_id
+        }
+
+    arxiv_data = arxiv_client.search_by_id(arxiv_id)
     # 持久化
     file_path, arxiv_id = download(url)
     arxiv_data.file_path = file_path
@@ -76,48 +96,34 @@ def trans(title: str, abstract: str, arxiv_id: str) -> str:
         return arxiv_data.title_abstract_cn
 
     content = f'## {title}\n{abstract}'
-    translated = deepseek.translate_en_zh(content)
+    translated = trans_agent.run(content)
     arxiv_data.title_abstract_cn = translated
     arxiv_data.save_to_json()
     return translated
 
 
-def chat_not_kimi(index: int, file_id: str, file_path: str, arxiv_id: str) -> (str, str):
+def answer_pdf(index: int, file_id: str, arxiv_id: str) -> (str, str):
     arxiv_data = get_data_from_arxiv_id(arxiv_id)
     if arxiv_data is None:
         return '系统异常'
+
     if arxiv_data.content is None or len(arxiv_data.content) == 0:
-        file_content = utils.read_pdf(arxiv_data.file_path)
-        arxiv_data.content = file_content
-        arxiv_data.save_to_json()
-    question = paper_questions[index]
-    if arxiv_data.faq is not None and arxiv_data.faq.get(question):
-        return file_id, arxiv_data.faq.get(question)
-    answer = deepseek.chat_pdf(question, arxiv_data.content)
-    print(answer)
-    arxiv_data.faq[question] = answer
-    arxiv_data.save_to_json()
-    return file_id, answer
-
-
-def chat_by_kimi(index: int, file_id: str, file_path: str, arxiv_id: str) -> (str, str):
-    arxiv_data = get_data_from_arxiv_id(arxiv_id)
-    if arxiv_data is None:
-        return '系统异常'
-
-    if file_id is None or file_id == '':
-        file_id = kimi.upload_file(file_path)
-        file_content = kimi.extract_file(file_id)
-        arxiv_data.file_id = file_id
-        arxiv_data.content = file_content
-        arxiv_data.save_to_json()
+        if isinstance(current_llm, KimiLlm):
+            file_id = current_llm.upload_file(arxiv_data.file_path)
+            file_content = current_llm.extract_file(file_id)
+            arxiv_data.file_id = file_id
+            arxiv_data.content = file_content
+            arxiv_data.save_to_json()
+        else:
+            file_content = utils.read_pdf(arxiv_data.file_path)
+            arxiv_data.content = file_content
+            arxiv_data.save_to_json()
 
     question = paper_questions[index]
     if arxiv_data.faq is not None and arxiv_data.faq.get(question):
         return file_id, arxiv_data.faq.get(question)
 
-    answer = kimi.chat_pdf(question, arxiv_data.content)
-    print(answer)
+    answer = current_llm.chat_pdf(question, arxiv_data.content)
     arxiv_data.faq[question] = answer
     arxiv_data.save_to_json()
     return file_id, answer
@@ -131,6 +137,25 @@ def create_files_dataframe(files):
     }
     df = pd.DataFrame(data)
     return df
+
+
+def export_md(arxiv_id: str):
+    arxiv_data = get_data_from_arxiv_id(arxiv_id)
+    path = utils.arxiv_dir_path(arxiv_id)
+    with open(md_template_path, 'r', encoding='utf-8') as f:
+        template = f.read()
+
+
+    template_format = template.format(title=arxiv_data.title, abstract=arxiv_data.abstract, title_abstract_cn=arxiv_data.title_abstract_cn)
+    faq = arxiv_data.faq
+    for key, value in faq.items():
+        item = '### ' + key + '\n' + value + '\n\n'
+        template_format += item
+
+    with open(path + os.sep + arxiv_id + '.md', 'w', encoding='utf-8') as f:
+        f.write(template_format)
+
+    print('导出结束')
 
 
 # 定义主页
@@ -150,13 +175,17 @@ def home():
 
     if 'arxiv_id' not in st.session_state:
         st.session_state.arxiv_id = ""
-    if 'file_path' not in st.session_state:
-        st.session_state.file_path = ""
     if 'translated_abstract' not in st.session_state:
         st.session_state.translated_abstract = ""
 
     if 'file_id' not in st.session_state:
         st.session_state.file_id = ""
+
+    if 'generate_all' not in st.session_state:
+        st.session_state.generate_all = False
+
+    if 'generate_index' not in st.session_state:
+        st.session_state.generate_index = 0
 
     url = st.text_input("请输入网址", value=st.session_state.url, key="url_input")
 
@@ -166,7 +195,6 @@ def home():
             st.session_state.title = analysis_result['title']
             st.session_state.abstract = analysis_result['abstract']
             st.session_state.arxiv_id = analysis_result['arxiv_id']
-            st.session_state.file_path = analysis_result['file_path']
             st.rerun()
 
     if url != st.session_state.url:
@@ -193,7 +221,19 @@ def home():
                             unsafe_allow_html=True)
 
     with col2:
-        if st.session_state.url:
+        if st.session_state.title:
+            spinner_placeholder = st.empty()
+            b1, b2 = st.columns(2)
+            with b1:
+                if st.button("生成所有"):
+                    st.session_state.generate_all = True
+                    st.session_state.generate_index = 0
+            with b2:
+                if st.button("导出MD"):
+                    with st.spinner("导出中，请稍候..."):
+                        export_md(st.session_state.arxiv_id)
+                        st.rerun()
+
             for i, question in enumerate(paper_questions):
                 with st.form(key=f"form_{i}"):
                     st.markdown(f"**{question}**", unsafe_allow_html=True)
@@ -201,17 +241,29 @@ def home():
                     submitted = st.form_submit_button("生成")
                     if submitted:
                         with st.spinner("生成中，请稍候..."):
-                            _, result = chat_not_kimi(i, st.session_state.file_id, st.session_state.file_path,
-                                                     st.session_state.arxiv_id)
+                            _, result = answer_pdf(i, st.session_state.file_id, st.session_state.arxiv_id)
                             st.session_state.responses[i] = result
                             st.rerun()
+
+    # 处理生成所有的问题
+    if st.session_state.generate_all and st.session_state.generate_index < len(paper_questions):
+        i = st.session_state.generate_index
+        with spinner_placeholder.container():
+            with st.spinner(f"正在生成问题 {i + 1}/{len(paper_questions)}..."):
+                _, result = answer_pdf(i, st.session_state.file_id, st.session_state.arxiv_id)
+                st.session_state.responses[i] = result
+                st.session_state.generate_index += 1
+                if st.session_state.generate_index >= len(paper_questions):
+                    st.session_state.generate_all = False
+        st.rerun()
 
 
 # 定义设置页面
 def settings():
-    st.markdown("<h1 style='text-align: center; font-size: 32px;'>Kimi文件管理</h1>", unsafe_allow_html=True)
+    st.markdown("<h1 style='text-align: center; font-size: 32px;'>Kimi文件管理(存在Kimi才可使用)</h1>",
+                unsafe_allow_html=True)
 
-    files = kimi.list_files()
+    files = current_llm.list_files()
     df = create_files_dataframe(files)
 
     # 显示文件表格
@@ -221,7 +273,7 @@ def settings():
         col2.write(row["FileName"])
         button_placeholder = col3.empty()
         if button_placeholder.button("删除", key=row["ID"]):
-            kimi.remove_file(row["ID"])
+            current_llm.remove_file(row["ID"])
             st.rerun()
